+++ package.json
@@ -11,7 +11,6 @@
     "ant-table-extensions": "^1.1.7",
     "antd": "^4.18.3",
     "moment": "^2.29.1",
-    "moment-js": "^1.1.15",
     "node-sass": "^7.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -22,10 +21,7 @@
     "react-scripts": "5.0.0",
     "react-yandex-maps": "^4.6.0",
     "redux": "^4.1.2",
-    "redux-localstorage-simple": "^2.4.1",
     "redux-persist": "^6.0.0",
-    "redux-saga": "^1.1.3",
-    "redux-thunk": "^2.4.1",
     "reduxjs-toolkit-persist": "^7.0.7",
     "web-vitals": "^2.1.3"
   },
