+++ src/Templates/ClientTemlates/InkorQilganMijozlarTemplate.js
