+++ src/components/tabs/BottomTabs.js
@@ -4,7 +4,6 @@ import { useSelector, useDispatch } from "react-redux";
 import { Tabs, Avatar, Badge } from "antd";
 import { removeTab, setCurrentPage } from "../../redux/tabs_reducer";
 import { useNavigate } from "react-router-dom";
-import iconArr from "../../assets/icons/icons.js"
 
 import {findIcon} from "../../assets/icons/icons"
 
@@ -61,6 +60,7 @@ const BottomTabs = () => {
           <TabPane
               tab={
                 <div className="site-footer__tab" onClick={() => onChange(i)}>
+                  {findIcon(pane?.icon)}
                   <span>{pane?.text}</span>
                 </div>
               }
