--- src/pageTemplates/MainTemplate.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import {MAIN_PAGE} from "../pages/pageConstants/PageTypes";
-import {MAIN_PATH} from "../pages/pageConstants/PageRoutes";
-import {MainIcon} from "../assets/icons/icons";
-
-const MainTemplate = {
-	text: "Asosiy Oyna",
-	path: MAIN_PATH,
-	icon: "MainIcon",
-	type: MAIN_PAGE,
-	// sections: [],
-};
-
-export default MainTemplate;
